feat: More safer minimum FS disruption mode
Before this commit, this operating mode performed like this:

1.  Remove the original file, preserving the handle.
2.  Create the destination file
3.  Copy/clone the contents from the original
    (which may take some time).

This commit now uses `O_TMPFILE` and performs like this:

1.  Create the destination file without path.
2.  Copy/clone the contents from the original
    (which may take some time).
3.  Replace the original file path with the destination file.

Though it does not prevent all TOCTOU cases, the original file is
preserved longer than before and makes the operation relatively safer.

It also preserves the original file contents as long as possible.